+++ skued/time_series/__init__.py
@@ -9,4 +9,11 @@ in the context of ultrafast diffraction.
 from .fitting import biexponential, exponential
 from .nfft_routines import nfft, nfftfreq
 from .robust import mad
+from .selections import (
+    ArbitrarySelection,
+    DiskSelection,
+    RectSelection,
+    RingSelection,
+    Selection,
+)
 from .time_zero import register_time_shift, register_time_shifts